Switch integer to long to support unsigned ints

+++ src/Service/SearchElastic/ElasticMapping.php
@@ -43,7 +43,7 @@ class ElasticMapping
     ];
 
     const INTEGER = [
-        'type' => 'integer',
+        'type' => 'long',
         'index' => true,
     ];
